backend/geop/requirements.txt
streamlit==1.31.0
openai==1.12.0
requests
python-dotenv
google-generativeai
flask
flask-cors